% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_editor.R
\name{get_active_file}
\alias{get_active_file}
\title{Get the file path of the active document}
\usage{
get_active_file()
}
\value{
A single string with the absolute path to the active document.
}
\description{
Get the file path of the active document
}
