% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_editor.R
\name{get_document}
\alias{get_document}
\title{Get the full contents of the active document}
\usage{
get_document()
}
\value{
A single string with the document source, lines joined by \code{"\\n"}.
}
\description{
Get the full contents of the active document
}
