% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_editor.R
\name{get_selection}
\alias{get_selection}
\title{Get the current editor selection}
\usage{
get_selection()
}
\value{
A list with \code{text} (the selected string) and \code{range} (the
\code{rstudioapi} document range).
}
\description{
Get the current editor selection
}
