% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_editor.R
\name{insert_at_cursor}
\alias{insert_at_cursor}
\title{Insert text at the current cursor position}
\usage{
insert_at_cursor(text)
}
\arguments{
\item{text}{The text to insert.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Insert text at the current cursor position
}
