% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_editor.R
\name{insert_at_end}
\alias{insert_at_end}
\title{Append text to the end of the active document}
\usage{
insert_at_end(text)
}
\arguments{
\item{text}{The text to append.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Append text to the end of the active document
}
