% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_editor.R
\name{replace_selection}
\alias{replace_selection}
\title{Replace the current selection with new text}
\usage{
replace_selection(text)
}
\arguments{
\item{text}{The replacement text.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Replace the current selection with new text
}
