% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_request.R
\name{user_message}
\alias{user_message}
\title{Build a user message object}
\usage{
user_message(content)
}
\arguments{
\item{content}{A string or a list of content blocks.}
}
\value{
A named list representing a user message.
}
\description{
Build a user message object
}
